changed CarParkKind to return BigDecimal

+++ src/CarParkKind.java
@@ -3,9 +3,10 @@ import java.math.BigDecimal;
 /**
  * Created by CM on 01/02/2018.
  */
-public enum CarParkKind {STAFF("staff", 0.00, 1.00, 0.00, 16.00),
+public enum CarParkKind {
+    STAFF("staff",  0, 1, 0, 16),
     STUDENT("student", 5.50, 0.25, 0.00, 0.00),
-    MANAGEMENT("management", 0.00, 1.00, 3.00, 0.00),
+    MANAGEMENT("management", 0.00, 1.00, 3, 0.00),
     VISITOR("visitor", 8.00, 0.50, 0.00, 0.00);
     private final String name;
     private final double free;
@@ -25,19 +26,19 @@ public enum CarParkKind {STAFF("staff", 0.00, 1.00, 0.00, 16.00),
         return this.name;
     }
 
-    public double getFree() {
-        return this.free;
+    public BigDecimal getFree() {
+        return BigDecimal.valueOf(this.free);
     }
 
-    public double getReduction() {
-        return this.reduction;
+    public BigDecimal getReduction() {
+        return BigDecimal.valueOf(this.reduction);
     }
 
-    public double getMinimum() {
-        return this.minimum;
+    public BigDecimal getMinimum() {
+        return BigDecimal.valueOf(this.minimum);
     }
 
-    public double getMaximum() {
-        return this.maximum;
+    public BigDecimal getMaximum() {
+        return BigDecimal.valueOf(this.maximum);
     }
 }
